Show/Read instances for Time now use mixed fractions
The Show and Read instances for Time now support mixed fractions:
`show` always outputs a mixed fraction, and `read` can read either
an improper fraction (the previous functionality) or a mixed
fraction.

The Show instance now has `showsPrec` explicitly implemented
instead of `show` as precedence could potentially matter with the
'+' in the middle.  Parentheses are added when the surrounding
precedence is greater than 6, the precedence of '+'.  For example,
3 * (7+2/5s).  This should not affect `seriesF` as that uses show,
and so a precdence of 0.

Resolves #61

Undo incorrect stylish-haskell formatting

Reran stylish-haskell on Test.Time.Units with the correct
.stylish-haskell.yaml

Update changelog with issue 61