[#110] Make benchmarks not buildable by default
Problem: our benchmarks depend on `time-units` and `tiempo` which are
rather dead libraries. Because of that we have to specify `allow-newer`
in Stack configurations and nightly resolver does not have `o-clock`
because it cares about benchmarks and these dependencies are not in
the resolver.

Solution: so we conditionally make benchmarks not buildable. One still
can build them by passing a new flag.