Update license to MPL-2.0
Problem: we consider MIT a bad license and internally we have decided
to change the license here to MPL-2.0. Also we have only one LICENSE
file which contradicts out best practices where for each file it should
be clear which license applies to it.

Solution: replace MIT license with MPL-2.0, make the repo REUSE
compliant.
Ideally we also want CI to check that `reuse lint` passes, it will
be done separately.