Drop tasty-hspec dependency
Problem: we write some tests using `hspec` and then convert them to
tasty trees using `tasty-hspec`. This conversion is not necessary,
but causes extra dependencies to be added to one test-suite.
Moreover, sometimes it's not 100% precise.

Solution: remove this dependency and group tests into trees ourselves.
It doesn't add any overhead, we are just using a slightly different
syntax.
We don't transitively depend on `hspec` anymore, but we need
`hspec-expectations` as a replacement to use things like `shouldThrow`.