Drop last bits of ghc<8.6 support
Problem: there are 4 major GHC versions newer than 8.4, so we can freely
disregard versions older than 8.6. But we still have CPP conditional
checking that GHC version is 8.6 or newer.

Solution: drop that CPP check.
Also require `base >= 4.12` since it's the version bundled with 8.6.1
and it seems to be the standard approach.