Fix doctests for GHC-9.2 and newer
Problem: there is the following change in GHC-9.2.1:
> GHCi’s :kind! command now expands through type synonyms
> in addition to type families.

We have some doctests where we expect things like `18 :% 11` in
output of `:kind!`, but `:%` is a type synonym.
Those tests pass before GHC-9.2.1 because type synonym is not
expanded, but they fail starting from GHC-9.2.1 because
`:%` is expanded into `'::%`.

Solution: use CPP to expect different results depending on GHC version.