[#124] Add lossless conversion to Fractional values
Problem: The function 'toNum' transforms an object of the type 'Time
timeunit' to a value of the type that satisfies the 'Num' constraint.
However, the former object can contain a fractional value inside, which
then needs to get floored before being converted to a Num value. This
could lead to a possible loss of precision.

Solution: create a separate function that converts the Time objects
losslessly to a value of the type that satisfies a stricter constraint:
'Fractional'. The lossless conversion can be done via 'fromRational'.