[Chore] Enable -fprint-redundant-promotion-ticks in Rational.hs
Problem: due to
https://github.com/ghc/ghc/commit/13d627bbd0bc3dd30d672de341aa7f471be0aa2c
we get different output in doctest for GHC 9.6 and 9.4.
So with 9.6 they started to fail.

Solution: force redundant promotion ticks for now. Add a comment that
we should remove it when we stop supporting older GHCs.
Another option would be to use CPP, but we already have 2 cases
and supporting 3 cases sounds too verbose.