[Chore] Build haddock only for the library
Problem: Haddock CI step tries to build haddock for everything and
fails to do it for doctest for an obscure reason.

Solution: we don't really care about Haddock for anything besides
the library, so the easiest solution seems to be to build Haddock only
for the library, which is the default.